+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## 1.0.0
 
 ### Added
+- `Mesh.Sphere(double radius, int divisions)`
+- `Material.EmissiveTexture`
+- `Material.EmissiveFactor`
 
 ### Changed
 - Remove ``removeCutEdges` from `AdaptiveGrid.SubtractBox` and always remove cut parts of intersected edges.